+++ test/libsolidity/syntaxTests/abstract/abstract_only.sol
@@ -1,0 +1,4 @@
+abstract
+// ----
+// ParserError: (9-9): Expected keyword "contract", "interface" or "library".
+// ParserError: (9-9): Expected identifier but got end of source